fix type of length passed to speex_header_to_packet() Bug reported by Jean-Marc Valin:
<jmworx> kfish: I think I found a bug in sweep with the way it uses speex_header_to_packet()
<jmworx> basically, the last argument is an int* and you pass a long*. Works fine on 32-bit archs (and maybe some 64-bit ones), but eventually you'll run into problems.
<jmworx> In any case, the fix is *probably* to just use an int for the size argument. You just need to make sure it won't break something else (haven't checked).
<jmworx> speexenc was passing ogg_packet.size directly, which was wrong. You did things sort of the right way (declare the size separately), but just didn't use the right type.


git-svn-id: http://svn.metadecks.org/sweep/trunk@439 e2d53364-18fc-0310-981e-b60a84739af0

+++ src/file_speex.c
@@ -796,7 +796,7 @@ speex_sample_save_thread (sw_op_instance * inst)
   /* write header */
 
   {
-    long bytes = op.bytes;
+    int bytes = op.bytes;
     op.packet = (unsigned char *)
       speex_header_to_packet (&header, &bytes);
     op.bytes = bytes;
